+++ src/fileHandlers/index.ts
@@ -1,5 +1,6 @@
 export * from './transfer';
 export * from './getSize';
+export * from './createArchive';
 export * from './remove';
 export * from './diff';
 export * from './rename';